R OBS Websocket Class

Create a websocket client to connect to OBS and send commands. The WebSocket server (Use "Tools" menu in OBS) must be activated.

Author

FML

Super class

websocket::WebSocket -> OBSWebSocket

Public fields

obsWebSocketVersion

Used to store the OBS websocket version of the server

challenge

Used to store the challenge sent by the OBS websocket during the connection and used for the identification

salt

Used to store the salt sent by the OBS websocket during the connection and used for the identification

outlog

Used to store the requests and responses to and from the OBS websocket server

Methods

Public methods

OBSWebSocket$clone()

Inherited methods

websocket::WebSocket$clearAccessLogChannels()

websocket::WebSocket$clearErrorLogChannels()

websocket::WebSocket$close()

websocket::WebSocket$connect()

websocket::WebSocket$initialize()

websocket::WebSocket$onClose()

websocket::WebSocket$onError()

websocket::WebSocket$onMessage()

websocket::WebSocket$onOpen()

websocket::WebSocket$protocol()

websocket::WebSocket$readyState()

websocket::WebSocket$send()

websocket::WebSocket$setAccessLogChannels()

websocket::WebSocket$setErrorLogChannels()

Method clone()

The objects of this class are cloneable with this method.

Usage

OBSWebSocket$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

myobs <- createOBSWebsocket()
#> ws://127.0.0.1:4455
connectToOBS(obs = myobs)
#> <OBSWebSocket>
#>   Inherits from: <WebSocket>
#>   Public:
#>     challenge: NULL
#>     clearAccessLogChannels: function (channels = c("all")) 
#>     clearErrorLogChannels: function (channels = c("all")) 
#>     clone: function (deep = FALSE) 
#>     close: function (code = 1000L, reason = "") 
#>     connect: function () 
#>     initialize: function (url, protocols = character(0), headers = NULL, autoConnect = TRUE, 
#>     obsWebSocketVersion: Unknown
#>     onClose: function (callback) 
#>     onError: function (callback) 
#>     onMessage: function (callback) 
#>     onOpen: function (callback) 
#>     outlog: data.frame
#>     protocol: function () 
#>     readyState: function () 
#>     salt: NULL
#>     send: function (msg) 
#>     setAccessLogChannels: function (channels = c("all")) 
#>     setErrorLogChannels: function (channels = c("all")) 
#>   Private:
#>     accessLogChannelValues: none connect disconnect control frame_header frame_paylo ...
#>     accessLogChannels: function (channels, stompValue) 
#>     callbacks: environment
#>     errorLogChannelValues: none devel library info warn rerror fatal all
#>     errorLogChannels: function (channels, stompValue) 
#>     getInvoker: function (eventName) 
#>     pendingConnect: FALSE
#>     wsObj: externalptr
identifyToOBS(obs = myobs, eventSub = "33", password = "OBSwebsocketServerPassword")
#> Error: invalid state
disconnectFromOBS(myobs)